projects/prescription/requirements.txt
django
djangorestframework
psycopg2-binary
drf_yasg
pika